change(migrations): integer to unsignedBigInteger

+++ Server/database/migrations/2024_10_27_180241_create_degrees_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('degrees', function (Blueprint $table) {
             $table->id();
             $table->string('degree_id')->unique();
-            $table->integer('school_id');
+            $table->unsignedBigInteger('school_id');
             $table->string('degree_name');
             $table->integer('degree_minimun_credits');
